\docType{methods}
\name{getAmpliconNames}
\alias{getAmpliconNames}
\alias{getAmpliconNames,AmpliconViews}
\alias{getAmpliconNames,AmpliconViews-method}
\title{get amplicon Names from AmpliconView object}
\usage{
getAmpliconNames(x)

\S4method{getAmpliconNames}{AmpliconViews}(x)
}
\arguments{
  \item{x}{an \code{\link{AmpliconViews}} object}
}
\value{
a character vector
}
\description{
get amplicon Names from AmpliconView object
}

