\docType{methods}
\name{getCoverage}
\alias{getCoverage}
\alias{getCoverage,AmpliconViews,character,character}
\alias{getCoverage,AmpliconViews,character,character-method}
\title{get coverage for bases in an amplicon}
\usage{
getCoverage(x,sampleName,ampliconName)

\S4method{getCoverage}{AmpliconViews,character,character}(x, sampleName,
  ampliconName)
}
\arguments{
  \item{x}{an \code{\link{AmpliconViews}} object}

  \item{sampleName}{name of the sample}

  \item{ampliconName}{name of the amplicon}
}
\description{
get coverage for bases in an amplicon
}

