\docType{methods}
\name{getExampleMethMat}
\alias{getExampleMethMat}
\alias{getExampleMethMat,AmpliconViews,character,character}
\alias{getExampleMethMat,AmpliconViews,character,character-method}
\title{get example methylation matrix
the methylation matrix of an ampliconBiSeq experiment}
\usage{
getExampleMethMat(x,sampleName,ampliconName)

\S4method{getExampleMethMat}{AmpliconViews,character,character}(x, sampleName,
  ampliconName)
}
\arguments{
  \item{x}{an \code{\link{AmpliconViews}} object}

  \item{sampleName}{name of the sample}

  \item{ampliconName}{name of the amplicon}
}
\description{
get example methylation matrix the methylation matrix of an
ampliconBiSeq experiment
}

