\docType{methods}
\name{show,AmpliconViews-method}
\alias{show,AmpliconViews}
\alias{show,AmpliconViews-method}
\title{show method for methylKit classes}
\usage{
\S4method{show}{AmpliconViews}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
The show method works for
\code{methylRaw},\code{methylRawList}, \code{methylBase}
and \code{methylDiff} objects
}
\examples{
data(ampViewEx)
show(ampViewEx)
}

